feat: move over to using the "mbwatch" group for watchers
Now mbwatch will look for a group called "mbwatch" to find what channels and
mailboxes to watch. This allows the user to be move specific in what they are
watching. You don't always need to watch the "INBOX" nor do you have to watch
all of the channels you have setup in with mbsync.

The config will need to be defined in the `Channels` prop on the group. It
needs to be a comma separated of channels and mailboxes. This is an example of
the config.

```
Group mbwatch
Channels work:INBOX, work:Archive, personal:INBOX
```